fixing deps and tests
