+++ SingularityService/pom.xml
@@ -131,6 +131,21 @@
     </dependency>
 
     <dependency>
+      <groupId>io.reactivex</groupId>
+      <artifactId>rxjava</artifactId>
+    </dependency>
+
+    <dependency>
+      <groupId>com.mesosphere.mesos.rx.java</groupId>
+      <artifactId>mesos-rxjava-client</artifactId>
+    </dependency>
+
+    <dependency>
+      <groupId>com.mesosphere.mesos.rx.java</groupId>
+      <artifactId>mesos-rxjava-util</artifactId>
+    </dependency>
+
+    <dependency>
       <groupId>org.apache.commons</groupId>
       <artifactId>commons-lang3</artifactId>
     </dependency>
