+++ pom.xml
@@ -178,6 +178,24 @@
       </dependency>
 
       <dependency>
+        <groupId>io.reactivex</groupId>
+        <artifactId>rxjava</artifactId>
+        <version>1.1.8</version>
+      </dependency>
+
+      <dependency>
+        <groupId>com.mesosphere.mesos.rx.java</groupId>
+        <artifactId>mesos-rxjava-client</artifactId>
+        <version>0.1.2</version>
+      </dependency>
+
+      <dependency>
+        <groupId>com.mesosphere.mesos.rx.java</groupId>
+        <artifactId>mesos-rxjava-util</artifactId>
+        <version>0.1.2</version>
+      </dependency>
+
+      <dependency>
         <groupId>com.spotify</groupId>
         <artifactId>docker-client</artifactId>
         <version>8.4.0</version>
